How many lives were lost due to inaction?
I've added plots to try to answer this for a few select countries.

I've used a simple exponential model to extrapolate the number of deaths in
upcoming weeks and use these numbers to get a rough sense of the difference
between the outcome we're looking at now, and the outcome we would be looking
at, had the government acted a week earlier.